Moved ProductImageMap to tools package, moved resetting it to the presenter in case activity is destroyed
