+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/main/MainActivity.kt
@@ -33,7 +33,6 @@ import com.woocommerce.android.ui.notifications.NotifsListFragment
 import com.woocommerce.android.ui.orders.OrderListFragment
 import com.woocommerce.android.ui.prefs.AppSettingsActivity
 import com.woocommerce.android.ui.sitepicker.SitePickerActivity
-import com.woocommerce.android.util.ProductImageMap
 import com.woocommerce.android.util.WooAnimUtils
 import com.woocommerce.android.util.WooAnimUtils.Duration
 import com.woocommerce.android.widgets.AppRatingDialog
@@ -80,7 +79,6 @@ class MainActivity : AppCompatActivity(),
     @Inject lateinit var presenter: MainContract.Presenter
     @Inject lateinit var loginAnalyticsListener: LoginAnalyticsListener
     @Inject lateinit var selectedSite: SelectedSite
-    @Inject lateinit var productImageMap: ProductImageMap
 
     private var isBottomNavShowing = true
     private lateinit var bottomNavView: MainNavigationView
@@ -278,7 +276,6 @@ class MainActivity : AppCompatActivity(),
             updateActivityTitle()
             refreshDashboard(true)
         }
-        productImageMap.reset()
     }
 
     private fun hasMagicLinkLoginIntent(): Boolean {
