+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/main/MainPresenter.kt
@@ -4,8 +4,9 @@ import com.woocommerce.android.annotations.OpenClassOnDebug
 import com.woocommerce.android.network.ConnectionChangeReceiver
 import com.woocommerce.android.network.ConnectionChangeReceiver.ConnectionChangeEvent
 import com.woocommerce.android.push.NotificationHandler.NotificationsUnseenChangeEvent
+import com.woocommerce.android.tools.ProductImageMap
+import com.woocommerce.android.tools.ProductImageMap.RequestFetchProductEvent
 import com.woocommerce.android.tools.SelectedSite.SelectedSiteChangedEvent
-import com.woocommerce.android.util.ProductImageMap.RequestFetchProductEvent
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
 import org.wordpress.android.fluxc.Dispatcher
@@ -33,7 +34,8 @@ class MainPresenter @Inject constructor(
     private val accountStore: AccountStore,
     private val siteStore: SiteStore,
     private val wooCommerceStore: WooCommerceStore,
-    private val notificationStore: NotificationStore
+    private val notificationStore: NotificationStore,
+    private val productImageMap: ProductImageMap
 ) : MainContract.Presenter {
     private var mainView: MainContract.View? = null
 
@@ -138,6 +140,7 @@ class MainPresenter @Inject constructor(
     @Subscribe(threadMode = ThreadMode.MAIN)
     fun onEventMainThread(event: SelectedSiteChangedEvent) {
         mainView?.resetSelectedSite()
+        productImageMap.reset()
 
         // Fetch a fresh list of order status options
         dispatcher.dispatch(WCOrderActionBuilder
